requirements.txt
amqp==5.0.9
asgiref==3.7.0
bandwidth-sdk==13.1.1
billiard==3.6.4.0
boto3==1.20.41
botocore==1.23.41
CacheControl==0.12.10
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.6
cron-descriptor==1.4.5
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==5.0
django-allauth==0.47.0
django-celery-beat==2.7.0
django-celery-results==2.2.0
django-ckeditor==6.4.1
django-compat==1.0.15
django-countries==7.2.1
django-crispy-forms==1.13.0
-e git+https://github.com/mjs7231/django-dbbackup.git@2ef053cda8078817ad6872cb3719e38de3cfdaeb#egg=django_dbbackup
django-debug-toolbar==4.2.0
django-extensions==3.1.5
django-filter==22.1
django-hijack==3.2.1
django-htmx==1.6.0
django-import-export==2.7.1
django-js-asset==2.0.0
django-recaptcha==3.0.0
django-ses==2.3.1
django-smtp-ssl==1.0
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
enum34==1.1.10
et-xmlfile==1.1.0
future==0.18.2
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
ijson==3.1.4
importlib-metadata==4.12.0
jmespath==0.10.0
jsonpickle==0.9.6
kombu==5.2.3
lxml==4.7.1
Markdown==3.4.1
MarkupPy==1.14
msgpack==1.0.3
numpy==1.21.5
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
prompt-toolkit==3.0.27
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.1
PyJWT==2.3.0
pyparsing==3.0.7
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rest-framework-simplejwt==0.0.2
rsa==4.8
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
stripe==2.64.0
tablib==3.1.0
tld==0.13
typing_extensions==4.3.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
sentry-sdk==1.32.0
